Extract coerceScalarFieldsToParsed utility
